WEB-2073: Improve Voucher activation logic.
